In huge levels, faces weren't rendered if multithreading was enabled and standard lighting was chosen.
git-svn-id: svn://svn.code.sf.net/p/d2x-xl/code/trunk@4931 01821946-851b-0410-b7f4-f1477c70e57a

+++ include/descent.h
@@ -1512,8 +1512,8 @@ class CFaceListIndex {
 		CShortArray				roots; // [MAX_WALL_TEXTURES * 3];
 		CShortArray				tails; // [MAX_WALL_TEXTURES * 3];
 		CShortArray				usedKeys; // [MAX_WALL_TEXTURES * 3];
-		short						nUsedFaces;
-		short						nUsedKeys;
+		int						nUsedFaces;
+		int						nUsedKeys;
 	public:
 		CFaceListIndex ();
 		void Init (void);
